requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0

# Database
duckdb==0.9.2
sqlalchemy==2.0.25

# Snowflake (optional - for cloud deployment)
# snowflake-connector-python==3.6.0

# Data Generation
faker==22.0.0

# Data Visualization
plotly==5.18.0
altair==5.2.0

# Data Quality (Optional)
# great-expectations==0.18.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3

